编写了商品保存到草稿、提交到商品表，并将保存操作记录的方法抽象出了，商品类目的二级查询实现了，查询操作记录也实现了，可能存在一些地方仍有不足，以及pom文件的依赖引入，组长记得做代码比对。

+++ pob-product/product-model/src/main/java/com/szml/pl/entity/ProductCategory.java
@@ -16,7 +16,7 @@ public class ProductCategory {
   private String categoryName;
   private Long parentId;
   private Long categoryLevel;
-  private Long status;
+  private Integer status;
   private String imgUrl;
   private Long leaf;
   private Long priority;
@@ -51,20 +51,20 @@ public class ProductCategory {
   }
 
 
-  public long getCategoryLevel() {
+  public Long getCategoryLevel() {
     return categoryLevel;
   }
 
-  public void setCategoryLevel(long categoryLevel) {
+  public void setCategoryLevel(Long categoryLevel) {
     this.categoryLevel = categoryLevel;
   }
 
 
-  public Long getStatus() {
+  public Integer getStatus() {
     return status;
   }
 
-  public void setStatus(Long status) {
+  public void setStatus(Integer status) {
     this.status = status;
   }
